% Generated by roxygen2 (4.0.2): do not edit by hand
\name{classify}
\alias{classify}
\title{\code{classify} classifies profiles according to speciate for simulation study}
\usage{
classify(profile, k = 10, sprofiles = speciate)
}
\arguments{
\item{profile}{Source apportionment profile matrix with number of columns equal to number of chemical constituents and number of rows equal to number of sources}

\item{k}{k corresponding to k nearest neighbors}

\item{sprofiles}{database of profiles from speciate}
}
\description{
\code{classify} classifies profiles according to speciate for simulation study
}

