% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nycdat.R, R/nycmdl.R
\docType{data}
\name{nydat}
\alias{nydat}
\alias{nydat}
\title{PM2.5 chemical constituent concentrations in New York City}
\format{A data frame with 174 rows and 32 variables}
\description{
A dataset of PM2.5 constituent concentrations from the
Queens College monitor in New York City from the New York
State Department of Environmental Conservation from
April 2001-December 2002.  The data contains concentrations 
for 30 chemical constituents of PM2.5.  The variables are:

A dataset of minimum detection limits (MDLs) corresponding to
PM2.5 constituent concentrations from the
Queens College monitor in New York City from the New York
State Department of Environmental Conservation from
April 2001-December 2002.  The data contains the maximum MDLs 
for 30 chemical constituents of PM2.5.  The variables are:
}
\details{
\itemize{
		\item Date. date of observation
		\item PM25. total mass PM2.5 concentration in mug/m^3
		\item aluminum. concentration of aluminum in mug/m^3
	...
}

\itemize{
		\item PM25. total mass PM2.5 concentration in mug/m^3
		\item aluminum. concentration of aluminum in mug/m^3
	...
}
}
\keyword{datasets}
