% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SIMapca}
\alias{SIMapca}
\title{Absolute Principal Component Analysis (APCA)}
\usage{
SIMapca(data, tots = NULL, nsources = NULL, adjust = NULL, mdl = NULL,
  cut = 1, type = "apca", mons = NULL, i = NA, print = F, k = 10,
  complete = F, ...)
}
\arguments{
\item{data}{data frame of daily constituent concentrations with date as first column}

\item{tots}{vector of total concentrations (total PM2.5) for each day.  If null, uses \code{rowSums(data)}}

\item{nsources}{number of sources.  If null, uses number of eigenvalues of the correlation matrix greater than one.}

\item{adjust}{method to adjust censored concentrations below minimum detection limits (MDLs).  If null, uses complete case data}

\item{mdl}{either a vector of mdls corresponding to each constituent or a matrix of mdls corresponding to each constituent and each day.}

\item{cut}{cutoff for eigenvalues.  Only used if nsources is NULL}

\item{type}{Source apportionment method (apca or mapca)}

\item{mons}{vector of monitors corresponding to data}

\item{i}{monitor number (for printing)}

\item{print}{whether to print monitors without temporal variability}
}
\description{
\code{simAPCA} applies APCA and extracts summary information for simulation study
}

