% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/speciate.R
\docType{data}
\name{speciate}
\alias{speciate}
\title{SPECIATE cleaned dataset of profiles}
\format{A data frame with 1744 rows and 23 variables}
\description{
A dataset of 1744 profiles rescaled from SPECIATE and named according to one of 7 categories: oil, dust, diesel, wood, vehicle, coal, metals.  Each profile (row) represents contributions from 23 chemical constituents of PM2.5.  Each row sums to 1.  The rownames can be made into a factor of labels for k nearest neighbors.
}
\keyword{datasets}

