% Generated by roxygen2 (4.0.1): do not edit by hand
\name{adjust}
\alias{adjust}
\alias{adjust.default}
\alias{print.adjust}
\title{adjust}
\usage{
adjust(x, ...)

\method{adjust}{default}(data, mdl, adjust = c("substitute", "exclude",
  "likelihood"), sub = 0.5, experc = 0.25, N = 10, burnin = 2, ...)

print.adjust(x)
}
\arguments{
\item{data}{data frame of daily constituent concentrations with date as first column}

\item{mdl}{vector or matrix of MDLs for each variable of data}

\item{adjust}{one of three methods.  'substitute' substitutes all censored constant proportion of the mdl (sub * mdl).  'exclude' excludes constituents with more than experc censored data.  'likelihood' multiply imputes censored data with a likelihood-based method.}

\item{sub}{proportion of MDL to substitute censored concentrations.  Default is 0.5.}

\item{experc}{If a constituent exceeds this value, using the exclude method, the constituent will be dropped from the analysis.}

\item{N}{number of draws from posterior for likelihood-based method}

\item{burnin}{number of samples to discard for likelihood-based method}
}
\description{
Adjusting data below the MDL
}
\details{
\code{adjust} adjusts censored concentrations below the MDL

These are functions to create complete PM2.5 constitutent
concentrations by substituting, excluding, or imputing censored
data. Works on dataframe where the first column is date and all subsequent
columns are concentrations of chemical constituents.
}
\examples{
data(nycdat)
data(nycmdl)
#fix data totals
pm <- nycdat$PM25
whPM <- which(colnames(nycdat) == "PM25")
nycdat <- nycdat[, -whPM]
whPM <- which(colnames(nycmdl) == "PM25")
nycmdl <- nycmdl[, -whPM]
adjust(nycdat, mdl = nycmdl, adjust = "substitute", sub = 1/sqrt(2))
adjust(nycdat, mdl = nycmdl, adjust = "exclude", experc = 0.2)
adjust(nycdat, mdl = nycmdl, adjust = "likelihood")
}

