% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/apca.R
\name{stdize1}
\alias{stdize1}
\title{\code{stdize1} Standardizes matrix so each column is mean 0 with variance 1}
\usage{
stdize1(data, i = "NA", print = F)
}
\arguments{
\item{data}{data frame of daily constituent concentrations with date as first column}

\item{i}{iteration number}

\item{print}{whether to print if variability of column is 0}
}
\description{
\code{stdize1} Standardizes matrix so each column is mean 0 with variance 1
}

